Building from Root no longer required for 'auto' inter-file padding
Building from Root no longer required to build/rebuild a disc of standard size. Previously, in order to build a disc to the default/standard GC disc size of ~1.36 GB, you would need to be building from a root folder with the "paddingBetweenFiles" set to "auto". Now you just need paddingBetweenFiles set to "auto" (only applicable in cases of rebuilding the disc).